_client_pygame/requirements.txt
pygame
python-socketio
